move libs to a libraries subfolder in src
