+++ src/cogs/cts.py
@@ -12,9 +12,9 @@ from discord import app_commands
 from discord.ext.commands import Cog  # pyright: ignore[reportMissingTypeStubs]
 from discord.utils import get
 
+from libraries.cts import get_lines, get_stop_times, get_stops
 from utils import ResponseType, response_constructor
 from utils.constants import GUILD_ID
-from utils.cts_api import get_lines, get_stop_times, get_stops
 from utils.errors import BaseError
 
 if TYPE_CHECKING:
