+++ src/libraries/openweathermap/models.py
